# Updates TinyGo to use Go 1.20 features