--- a/.github/workflows/commit.yaml
+++ b/.github/workflows/commit.yaml
@@ -25,7 +25,7 @@ jobs:
     strategy:
       matrix:
         go-version:  # Note: Go only supports 2 versions: https://go.dev/doc/devel/release#policy
-          - "1.19"
+          - "1.20"
 
     steps:
       - uses: actions/checkout@v3
@@ -50,17 +50,16 @@ jobs:
     strategy:
       matrix:
         go-version:  # Note: Go only supports 2 versions: https://go.dev/doc/devel/release#policy
-          - "1.19"
+          - "1.20"
         tinygo-version:  # Note: TinyGo only supports latest: https://github.com/tinygo-org/tinygo/releases
           - "0.27.0"  # Latest
 
     steps:
       - uses: actions/checkout@v3
 
-      - uses: actions/setup-go@v3
+      - uses: actions/setup-go@v4
         with:
           go-version: ${{ matrix.go-version }}
-          cache: true
 
       - name: "Set up TinyGo"
         run: |  # Installing via curl so commands are similar on OS/x