--- a/.github/workflows/testdata.yaml
+++ b/.github/workflows/testdata.yaml
@@ -27,17 +27,16 @@ jobs:
     strategy:
       matrix:
         go-version:  # Note: Go only supports 2 versions: https://go.dev/doc/devel/release#policy
-          - "1.19"
+          - "1.20"
         tinygo-version:  # Note: TinyGo only supports latest: https://github.com/tinygo-org/tinygo/releases
           - "0.27.0"  # Latest
 
     steps:
       - uses: actions/checkout@v3
 
-      - uses: actions/setup-go@v3
-        with:  # Not cache: true because we also cache go-build and golangci-lint
+      - uses: actions/setup-go@v4
+        with:
           go-version: ${{ matrix.go-version }}
-          cache: true
 
       - name: "Set up TinyGo"
         run: |  # Installing via curl so commands are similar on OS/x
@@ -48,7 +47,7 @@ jobs:
 
       - name: "Set up wat2wasm"
         run: |  # Needed for testdata. wabt includes wat2wasm.
-          wabt_version=1.0.32
+          wabt_version=1.0.33
           wabt_url=https://github.com/WebAssembly/wabt/releases/download/${wabt_version}/wabt-${wabt_version}-ubuntu.tar.gz
           curl -sSL ${wabt_url} | tar --strip-components 2 -C /usr/local/bin -xzf - wabt-${wabt_version}/bin/wat2wasm