--- a/.golangci.yml
+++ b/.golangci.yml
@@ -1,6 +1,6 @@
 # This is built using - https://golangci-lint.run/usage/configuration/
 run:
-  go: '1.19'
+  go: '1.20'
   modules-download-mode: readonly
   tests: false