--- a/internal/e2e/go.mod
+++ b/internal/e2e/go.mod
@@ -4,9 +4,9 @@ go 1.18
 
 require (
 	github.com/http-wasm/http-wasm-guest-tinygo v0.0.0
-	github.com/http-wasm/http-wasm-host-go v0.3.5
-	github.com/stretchr/testify v1.8.2
-	github.com/tetratelabs/wazero v1.0.0-rc.2
+	github.com/http-wasm/http-wasm-host-go v0.5.1
+	github.com/stretchr/testify v1.8.4
+	github.com/tetratelabs/wazero v1.2.0
 )
 
 require (