… examples/router/main.wasm
… examples/wasi/main.wasm
… internal/e2e/go.sum
… internal/test/testdata/bench/tinygo/add_header_value/main.wasm
… internal/test/testdata/bench/tinygo/get_header_names/main.wasm
… internal/test/testdata/bench/tinygo/get_header_values/main.wasm
… internal/test/testdata/bench/tinygo/get_uri/main.wasm
… internal/test/testdata/bench/tinygo/log/main.wasm
… internal/test/testdata/bench/tinygo/read_body/main.wasm
… internal/test/testdata/bench/tinygo/read_body_stream/main.wasm
… internal/test/testdata/bench/tinygo/remove_header/main.wasm
… internal/test/testdata/bench/tinygo/set_header_value/main.wasm
… internal/test/testdata/bench/tinygo/set_status_code/main.wasm
… internal/test/testdata/bench/tinygo/set_uri/main.wasm
… internal/test/testdata/bench/tinygo/stream_body/main.wasm
… internal/test/testdata/bench/tinygo/write_body/main.wasm
… internal/test/testdata/e2e/tinygo/features/main.wasm
… internal/test/testdata/e2e/tinygo/handle_response/main.wasm
… internal/test/testdata/e2e/tinygo/log/main.wasm