TinyGo 0.27 doesn't fully support Go 1.20, but it supports what we need. Particularly, `unsafe.SliceData` and `unsafe.StringData` were added to TinyGo 0.27 ahead of the more complete Go 1.20 version: 0.28 (not yet out).

Using the approach elaborated in wazero [1.2.0](https://github.com/tetratelabs/wazero/releases/tag/v1.2.0) TinyGo examples resulted in significant performance benefits, benefits far more than just the part around the runtime.

Here's the effective before/after
```bash
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/http-wasm/http-wasm-guest-tinygo/e2e
                                        │   old.txt   │              new.txt               │
                                        │   sec/op    │   sec/op     vs base               │
/example_router_host_response/TinyGo-12   636.8n ± 2%   609.5n ± 1%   -4.28% (p=0.002 n=6)
/example_wasi/TinyGo-12                   9.191µ ± 2%   7.066µ ± 1%  -23.12% (p=0.002 n=6)
/example_router_wasm_response/TinyGo-12   855.3n ± 1%   677.2n ± 2%  -20.82% (p=0.002 n=6)
```

The only constraint added, is that users need to compile with Go 1.20, which they would anyway in a couple months when Go 1.21 would make that the bases version anyway. As this is an early project, the tradeoff seems sound.

See https://github.com/tinygo-org/tinygo/commit/c43958972c3ffcd51e65414a346e53779edb9f97