@codefromthecrypt 2023-06-05
ok also re-ran the benchmark which verifies the gains were legit.. repasted results